First commit: test build of EEGLAB compiled version
Note: as of 13 Oct 2021 this cannot currently use main branch of neurodocker, as matlab mcr is not available. This currently relies on the 20210924-matlabmcr branch - BUT also requires edits to this matlabmcr yaml to include recent matlab runtime versions. Note also, additional dependencies for mcr installation are included in build.sh (curl ca-certificates unzip)